Move code to create scan storages from cli to scanner
Move the code to create scan storages from `ScannerCommand` into the new
class `ScanStorages` in the scanner module to make it reusable.
